Move create_cellid_da to mf6 wel adapter module
